Fix bugs. Refactor and comment code. Implement start screen GUI.
Started with refactoring of methods and removing magic numbers to make the code a lot more readable. There were several bugs associated with timing of events, fixed these by checking timing events throughout the loop. The WiFi connect() method was not reconnecting, fixed this also. Removed unused library and definition declaration from the main.ino and button.h class. Need to test button class again in the main to reintroduce it, because it was not working this time. Fixed MQTT topic errors. Changed some of the MQTT topics in mqttTopics.h and parser() method to prepare for sprint 3. Next steps are to make classes (if and where possible).