Implement notifications. Fix timer. Refactoring.
The major objective of this commit was to address software quality issues (see issue #31 in git) with respect to the main.ino file. Although the main.ino met functional requirements for features in Sprint 1 ( and Sprint 2, it lacked modularity and structure. In this commit, I refactored a lot of the code so that it is broken up into smaller methods, more organized, and more commented. In particular, I have fixed timing issues (#36) in the code caused by extraneous Serial.print statements and delay() method calls. It is really important to note that these are blocking methods, during which the Arduino is incapable of processing anything else. By switching from delay() to nonBlockingDelay() - we are able to free the Arduino/Wio Device microprocessor to continue checking MQTT for new messages. This is a workaround because the ideal way to do this would have been to redesign the project and implement a State Machine design pattern, with external states, and internal sub-states. In the current design, we have to basically do loop() to cycle between external states, and the nonBlocking method to while loop() through the internal sub-states. Another problem was with the button (issue #23), which implemented an indefinite blocking delay - as a result, I had to reintroduce the checkButton() function into the Button class and implement the nonBlockingDelay while waiting -- mainly so that users do not miss important notifications which are probably more important than "stand up and stretch" prompts. Additionally, I refactored parsers and left room for new features to be implemented in Sprint 3, such as user defined mood settings (#10) (perhaps to play music on the speaker), and user defined time intervals to change event timing intervals (a simple parser will do). Lastly, I implemented default and user-defined notification messages (#18), as well as default and user-defined motivational messages (#13). This required overloading of some display methods to figure out how to best display these on a TFT that does not have dynamic text-wrapping. The Wio is not the best place to do more complicated processing. If Sprint 3 features are scrapped, this is sufficient for a version 1.0 of the project, and so I have prepared a header for this file. The environment dashboard on the Wio (issue #5) also seems to be responding appropriately now that the timing issues are solved -- will close the issue after merging. Next commit will attempt to create classes for Parser, and Display (issue #35). At this stage, I do not think a Connectivity class will be possible for MQTT because it is integral to the main.ino which functions as a controller in MVC architecture. The Actuator classes are finished (issue #34). The Sensors (#33) do not have functions other than one line readSensor() method calls from analog pin or sensor library, so I do not feel the need to make this class. The Timer class (issue #36) is integral to the control flow, and so, I am of the opinion that it belongs in the main for now.