Fix bug related to Strings becoming corrupted
Implemented char array message buffer in the main, and replaced the eventMessage attribute in Event.h class back to String. I was improperly using Strings, which apparently have dynamic memory in C++. It is therefore safe to create new String objects and the garbage collector in C++ automatically knows to destroy the ones not assigned references. See issue #41