Immutable messages and primitive runtime exception handling
Added immutable String objects to capture incoming topics and messages in MQTT callback. Added also runtime exception handling to the Sensor class, as an additional safety check when setting indicator colors. If the message is not an expected color, it will trigger a Serial print message about the runtime exception in the setIndicatorColor() method. This is an additional layer of safety to the protect the integrity and encapsulation of the Sensor class. With this bugfix, I hope that issue #41 will not be encountered again!