Remove unused Utils.
In the previous commit, I had tried and tested Utils.init(this) as a way to remove a runtime error related to MPAndroidChart. The runtime error persisted after doing what the exception tells us to do. A cursory search of the internet pointed out that this may be leftover from deprecated features more than 5 years ago. Therefore I removed Utils.init in the graph classes, but accidentally left one out. Since we are not using this, I am removing unused library import to Utils.