Add LED, const pins, modify standup event.
In preparation for Sprint 3 demo, I have added the mood light feature (#50) which turns on while waiting for the user to standup and stretch event. I have also made the pins as constants at the beginning of the main file, so they are all in one place. Also modified standup event to include app based standup triggers. Change the default standup, motivate, and notification intervals to 1 hour, 30 minutes, and 24 hours, respectively.