Modify MqttHandler to generate unique ClientIDs
This is a bugfix for issue #60. The MQTT broker needs connections from unique Client ID's. This commit allows the MqttHandler singleton class to generate unique client ID's. This will remove the hardcoded Client ID, and allow multiple developers to connect to the broker without having to change the hardcoded Client ID in their git directory.