Populate database with live sensor data
This commit implements feature #64, by populating the Firebase realtime database with live sensor data as it is received in the MainActivity class. It works by establishing a reference to an existing node in the database, which is determined by parsing the MQTT topic, and inserting a child node with unique key, and value (timestamp, sensor_value).