Implement user-defined timing intervals in Wio
This commit addresses issue #61, which concerns user-defined timing intervals for standUp prompts, motivational messages, and notifications on the Wio. I revived the MQTT topic called setTiming from Sprint 2. When a message of this topic comes from MQTT, parseTiming is called, and sorts incoming messages which are coded by their first character from 0 to 7. 0 is for default timing intervals. 1, 2, 3, are to modify standUp, motivate, notification events, respectively. A message from topic setTiming with content 13600000 will set the standUp interval to 3600000 milliseconds, which is 1 hour. 4, 5, 6 are to turn off standUp, motivate, notification events, respectively. 7 is to turn off all events. This is clearly explained in the comments. I am also modifying gitignore to allow for MQTT topics to be seen because the android app src code has a file for this as well, and the topics are no longer a secret as we are not using a public broker.