Refactor Wio Sensor class, parsers, to save prefs
This commit concerns issue #65. Previously, the Wio changed indicator colors only when it received MQTT messages (TFT_GREEN, TFT_ORANGE, TFT_RED) under topics TOPIC_SUB_TEMP, TOPIC_SUB_HUMID, and TOPIC_SUB_LIGHT. This was fine, but required the app to publish messages before every dashboard refresh on the Wio. In this commit, I have refactored the code so that we can now publish the temperature, humidity, and light level preferences to the Wio (such as temperature 25.0 degC, humidity 35.0 %, and light level 900.0 lx), rather than the indicator colors. These are then saved as attributes in the Sensor class. Sensor objects are now initialized with default preferences, and the indicator colors change automatically according to default settings or if the user changes their preferences via the app. This is nice, because if the app is turned off, it is no longer a problem, as the Wio will continue to update its indicator colors automatically anytime it reads new sensor values. In addition to this refactoring, I have also added comments and fixed spacing of comments in the Wio classes to make them more consistent (easier to read).