Remove blocking delay. Fix notification interval.
Removed the blocking delay associated with the button's delayUntilPressed method in notificationEventSequence and replaced it with the non-blocking version found in standUpEventSequence. Changed the default notification interval to the specified amount which is 6 hours, not 7 hours. Renamed button.h and button.cpp to Button.h and Button.cpp because class names should be capitalized.