Remove flicker from dashboard TFT updates.
Modified display class to remove the clear screen from the drawDashboard method, and to only erase sensor values that change. This means that the refresh only requires setting the sensor values that change to black and then white again, effectively removing the flicker when we want a responsive dashboard (0.5-1 second refresh intervals).