Implement CI for Arduino (Wio Terminal)
This commit concerns CI-CD, issue #89. After writing up the automated build wiki page, I worked on a CI pipeline for the Wio Terminal, based on an existing GitLab CI pipeline for the esp32 (Credits to Pasha Klimenkov). I did this in a parallel project in GitLab. This relies on setting 3 environment variables, one for MQTT connection secrets, one for WiFi secrets, and another for the USB port you will be uploading to if the runner is on your local machine. I unfortunately do not have any tests for this. Committing anyways to document learning. The CI pipeline works!