Change gradle version to latest
This commit pertains to CI-CD for the Android app (see issue #91). There was a problem with the gradle version while I was trying to go through and document the automated build process testerday. The automated build worked only after changing the gradle version to latest. The installation steps related to automated build is described in detail in the wiki page.